move StringInt to jsonutil package
