+++ util/jsonutil/stringInt.go
@@ -1,4 +1,4 @@
-package typeutil
+package jsonutil
 
 import (
 	"github.com/buger/jsonparser"
